Quixote/requirements.txt
asgiref==3.4.1
boto3==1.21.27
botocore==1.24.27
certifi==2021.10.8
charset-normalizer==2.0.9
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
decouple==0.0.7
dj-database-url==0.5.0
Django==3.2.8
django-crispy-forms==1.13.0
django-utils==0.0.2
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
MarkupSafe==2.0.1
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.0
Pillow==9.0.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==6.0.0
